pyproject.toml
[tool.poetry]
name = "misp_testing"
version = "0.1.0"
description = "Testing framework for MISP, Docker based"
authors = ["Raphaël Vinot <raphael@vinot.info>"]

[tool.poetry.dependencies]
python = "^3.7"
pymisp = "^2.4.122"
gitpython = "^3.1.0"
nose = "^1.3.7"

[tool.poetry.dev-dependencies]
ipython = "^7.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
